I) Wine-upstream: 29 merge requests
1. https://gitlab.winehq.org/wine/wine/-/merge_requests/8982: ole32: Fix the return value for StgOpenStorageOnILockBytes().
2. https://gitlab.winehq.org/wine/wine/-/merge_requests/8981: dwrite: Always initialize &#39;contours&#39; flag.
3. https://gitlab.winehq.org/wine/wine/-/merge_requests/8974: dmsynth: Don&#39;t force enable DMUS_EFFECT_REVERB.
4. https://gitlab.winehq.org/wine/wine/-/merge_requests/8973: winex11.drv: Remove a no longer valid assert().
5. https://gitlab.winehq.org/wine/wine/-/merge_requests/8971: make_unicode: Add some halfwidth mapping exceptions.
6. https://gitlab.winehq.org/wine/wine/-/merge_requests/8968: setupapi: Parse a 0x prefix in SetupGetBinaryField().
7. https://gitlab.winehq.org/wine/wine/-/merge_requests/8964: winebuild: Use .rdata section instead of .rodata on PE targets.
8. https://gitlab.winehq.org/wine/wine/-/merge_requests/8962: dbghelp: Fix pointer truncation issues when parsing 64-bit Mach-O binaries in a Wow64 process
9. https://gitlab.winehq.org/wine/wine/-/merge_requests/8956: nls: Update character tables to Unicode 17.0.0.
10. https://gitlab.winehq.org/wine/wine/-/merge_requests/8954: winecfg: Allow configuring default MIDI device.
11. https://gitlab.winehq.org/wine/wine/-/merge_requests/8950: winetest: Print Windows revision (UBR).
12. https://gitlab.winehq.org/wine/wine/-/merge_requests/8948: iphlpapi: Fix comparison functions.
13. https://gitlab.winehq.org/wine/wine/-/merge_requests/8947: Introduce a way to generate arbitrary binary files in cmd.exe&#39;s tests.
14. https://gitlab.winehq.org/wine/wine/-/merge_requests/8945: opcservices: IOpcPart::GetContentType takes LPWSTR, not BSTR.
15. https://gitlab.winehq.org/wine/wine/-/merge_requests/8944: winedbg: Reserve more memory for symbol value. (ASan).
16. https://gitlab.winehq.org/wine/wine/-/merge_requests/8943: ntdll: Fix a memory overflow.
17. https://gitlab.winehq.org/wine/wine/-/merge_requests/8941: nsi: Explicitly add loopback entries to ipv4 forward table on Linux.
18. https://gitlab.winehq.org/wine/wine/-/merge_requests/8938: gdiplus: Manually blend to white background if the device doesn&#39;t support alpha blending.
19. https://gitlab.winehq.org/wine/wine/-/merge_requests/8935: d3d8, d3d9: Flag some failing tests as todo_wine.
20. https://gitlab.winehq.org/wine/wine/-/merge_requests/8933: mf: Update the stream sink input type when a transform output type changes.
21. https://gitlab.winehq.org/wine/wine/-/merge_requests/8932: combobox: Don&#39;t update item height on WM_SETFONT for owner drawn CB.
22. https://gitlab.winehq.org/wine/wine/-/merge_requests/8930: Fix check for early IRP cancellation
23. https://gitlab.winehq.org/wine/wine/-/merge_requests/8928: msxml3/tests: add test for IXMLDOMElement_removeAttributeNode.
24. https://gitlab.winehq.org/wine/wine/-/merge_requests/8924: oleaut32: Fix UDT record block leak in VariantClear().
25. https://gitlab.winehq.org/wine/wine/-/merge_requests/8918: wineboot: Fix a memory leak in create_computer_name_keys.
26. https://gitlab.winehq.org/wine/wine/-/merge_requests/8917: gdi32: From Windows 2000 FixBrushOrgEx() is a NOP.
27. https://gitlab.winehq.org/wine/wine/-/merge_requests/8916: ntdll: Switch to CLOCK_BOOTTIME for monotonic counters on Linux.
28. https://gitlab.winehq.org/wine/wine/-/merge_requests/8915: d3dx10_33/tests: Enable tests.
29. https://gitlab.winehq.org/wine/wine/-/merge_requests/8914: msxml: IXMLDOMNamedNodeMap::getNamedItem to handle global namespace.

II) Wine-upstream: refreshed 2 libs

- ldap: refreshed
- xml2: refreshed

III) Wine-staging (Wine 10.15): applied 'mf_http_support' patches set

- re-worked on some patches

+++ wine-red-rose-custom/dlls/dbghelp/image_private.h
@@ -124,8 +124,8 @@ struct image_file_map
         } elf;
         struct macho_file_map
         {
-            size_t                      segs_size;
-            size_t                      segs_start;
+            UINT64                      segs_size;
+            UINT64                      segs_start;
             HANDLE                      handle;
             struct image_file_map*      dsym;   /* the debug symbols file associated with this one */
             size_t                      header_size; /* size of real header in file */
